rename LifecycleConfiguration to S3LifecycleConfiguration
