+++ hadoop-ozone/s3gateway/src/test/java/org/apache/hadoop/ozone/s3/endpoint/TestS3LifecycleConfigurationDelete.java
@@ -18,7 +18,7 @@
 package org.apache.hadoop.ozone.s3.endpoint;
 
 import static java.net.HttpURLConnection.HTTP_NOT_FOUND;
-import static java.net.HttpURLConnection.HTTP_OK;
+import static java.net.HttpURLConnection.HTTP_NO_CONTENT;
 import static org.apache.hadoop.ozone.s3.exception.S3ErrorTable.NO_SUCH_LIFECYCLE_CONFIGURATION;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
@@ -35,10 +35,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 /**
- * Testing for GetBucketLifecycleConfiguration.
+ * Testing for DeleteBucketLifecycleConfiguration.
  */
-public class TestLifecycleConfigurationGet {
-
+public class TestS3LifecycleConfigurationDelete {
   private OzoneClient clientStub;
   private BucketEndpoint bucketEndpoint;
 
@@ -53,10 +52,29 @@ public class TestLifecycleConfigurationGet {
   }
 
   @Test
-  public void testGetNonExistentLifecycleConfiguration()
+  public void testDeleteNonExistentLifecycleConfiguration()
       throws Exception {
     try {
-      bucketEndpoint.get("bucket1", null, null, null, 0, null, null,
+      bucketEndpoint.delete("bucket1", "");
+      fail();
+    } catch (OS3Exception ex) {
+      assertEquals(HTTP_NOT_FOUND, ex.getHttpCode());
+      assertEquals(NO_SUCH_LIFECYCLE_CONFIGURATION.getCode(),
+              ex.getCode());
+    }
+  }
+
+  @Test
+  public void testDeleteLifecycleConfiguration() throws Exception {
+    String bucketName = "bucket1";
+    bucketEndpoint.put(bucketName, null, "", null, getBody());
+    Response r = bucketEndpoint.delete(bucketName, "");
+
+    assertEquals(HTTP_NO_CONTENT, r.getStatus());
+
+    try {
+      // Make sure it was deleted.
+      bucketEndpoint.get(bucketName, null, null, null, 0, null, null,
           null, null, null, null, null, 0, "", null);
       fail();
     } catch (OS3Exception ex) {
@@ -64,24 +82,6 @@ public class TestLifecycleConfigurationGet {
       assertEquals(NO_SUCH_LIFECYCLE_CONFIGURATION.getCode(),
           ex.getCode());
     }
-  }
-
-  @Test
-  public void testGetLifecycleConfiguration() throws Exception {
-    String bucketName = "bucket1";
-    bucketEndpoint.put(bucketName, null, "", null, getBody());
-    Response r = bucketEndpoint.get(bucketName, null, null, null, 0, null, null,
-        null, null, null, null, null, 0, "", null);
-
-    assertEquals(HTTP_OK, r.getStatus());
-    LifecycleConfiguration lcc =
-        (LifecycleConfiguration) r.getEntity();
-    assertEquals("remove logs after 30 days",
-        lcc.getRules().get(0).getId());
-    assertEquals("prefix/", lcc.getRules().get(0).getPrefix());
-    assertEquals("Enabled", lcc.getRules().get(0).getStatus());
-    assertEquals(30,
-        lcc.getRules().get(0).getExpiration().getDays().intValue());
   }
 
   private static InputStream getBody() {
